major refactoring: fixed css naming to follow convention, re-organised files structure, no longer using css modules for styling
