--- src/App.css
@@ -1,18 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.App button {
-  padding: 16px;
-  font: inherit;
-  color: white;
-  cursor: pointer;
-}
-
-.App button.Red {
-  background-color: red;
-}
-
-.App button:hover {
-  color: black;
-}
