--- src/components/AboutMe/AboutMe.css
@@ -1,10 +1,0 @@
-.AboutMe {
-  background-color: #ffffff;
-  color: #666666;
-  text-align: center;
-  line-height: 1.5em;
-  font-size: 1.5em;
-  padding-bottom: 20px;
-  font-weight: normal;
-
-  }
