--- src/components/AboutMe/AboutMe.jsx
@@ -1,19 +1,0 @@
-import React from 'react'
-import classes from './AboutMe.css'
-import Content from '../Layout/Content/content'
-
-const AboutMe = (props) => (
-  <React.Fragment>
-    <Content>
-      <div className={classes.AboutMe}>
-      <h1>ABOUT ME</h1>
-      <br/>
-      <p>I'm Maya. I love to play guitar and sing, and I have a cat.</p>
-          <p>I love building digital products and creating intuitive websites with a focus on aesthetic offering a great user experience. Last year I decided to take the plunge by quitting my steady job and enrolled in the Enspiral Dev Academy. It's an 18 week, 60+ hours per week web development bootcamp with intense learning to become a full-stack web developer. 
-      </p>
-      </div>
-    </Content>
-  </React.Fragment>
-)
-
-export default AboutMe
