+++ src/components/Layout/Content/content.js
@@ -2,7 +2,6 @@ import React from 'react'
 import PropTypes from 'prop-types';
 import { withStyles } from '@material-ui/core/styles';
 import Grid from '@material-ui/core/Grid';
-import classes from '../../../index.css'
 
 const styles = theme => ({
   root: {
@@ -13,7 +12,7 @@ const styles = theme => ({
 
 const Content = (props) => {
   return (
-    <Grid container spacing={24} className={classes.Content}>
+    <Grid container spacing={24} className='content'>
       <Grid item xs={3} />
         <Grid item xs={6}>
           {props.children}
