--- src/index.css
@@ -1,18 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-body {
-  font-family: "Roboto", sans-serif;
-  font-weight: 200;
-  font-size: 16px;
-  color: #ffffff;
-  line-height: 1.5em;
-  
-}
-
-.Content {
-  padding-top: 15vh;
-}
