+++ src/index.js
@@ -1,7 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
 import App from './App';
 import {BrowserRouter} from 'react-router-dom'
-const app = <BrowserRouter><App /></BrowserRouter>
+
+const app = <BrowserRouter>
+              <App />
+            </BrowserRouter>
+            
 ReactDOM.render(app, document.getElementById('root'));
